+++ CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased](https://github.com/edtr-io/edtr-io/compare/0.3.0..HEAD)
 
+### Changed
+
+- **plugin-spoiler**. Remove input element in render mode
+
 ### Fixed
 
-- Don't dispatch selection changes in text plugin anymore
+- **plugin-text**. Don't dispatch selection changes anymore
 
 ## [0.3.0](https://github.com/edtr-io/edtr-io/compare/0.2.1..0.3.0) - March 8, 2019